requirements.txt
# Core dependency — required for benchmark orchestration (benchmark_catalog.py, policy_utils.py, etc.)
pyyaml>=6.0

# Required for corpus ELF analysis helpers (generate_default_policies.py, runner/libs/inventory.py)
pyelftools>=0.29

# Required for figure generation and statistical analysis (generate_figures.py)
matplotlib>=3.7
numpy>=1.24
scipy>=1.10
